Removed support for Insider Preview Builds (read desc)
Support for insider preview builds has been removed. To restore support, replace the 1 in the line 

If ($imageInfo.SPBuild -notmatch "^1$")

with the corresponding SPBuild number of the insider build. like (1000, 5000, 5001, 1001 etc..)

Also, in:

26100 { Download-Files "https://....

Replace build 26100 with the insider preview build number. (e.g. 26742)

There are no plans to restore support for Windows Insider preview builds in the future.